refactor: implement lazy load for MarkdownViewer component to reduce main bundle size
