+++ StatusPage/src/Pages/Overview/Overview.tsx
@@ -48,7 +48,7 @@ import EmptyState from 'CommonUI/src/Components/EmptyState/EmptyState';
 import IconProp from 'Common/Types/Icon/IconProp';
 import API from '../../Utils/API';
 import StatusPage from 'Model/Models/StatusPage';
-import MarkdownViewer from 'CommonUI/src/Components/Markdown.tsx/MarkdownViewer';
+import MarkdownViewer from 'CommonUI/src/Components/Markdown.tsx/LazyMarkdownViewer';
 import StatusPageUtil from '../../Utils/StatusPage';
 import HTTPErrorResponse from 'Common/Types/API/HTTPErrorResponse';
 import { STATUS_PAGE_API_URL } from '../../Utils/Config';
